+++ source/Messaging.Domain/Actors/MarketRoles.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Messaging.Application.OutgoingMessages.Common;
+namespace Messaging.Domain.Actors;
 
 public static class MarketRoles
 {
