+++ source/Messaging.IntegrationTests/Application/Transactions/MoveIn/Notifications/NotifyCurrentEnergySupplierHandlerTests.cs
@@ -19,6 +19,7 @@ using Messaging.Application.Configuration.DataAccess;
 using Messaging.Application.OutgoingMessages;
 using Messaging.Application.OutgoingMessages.Common;
 using Messaging.Application.Transactions.MoveIn.Notifications;
+using Messaging.Domain.Actors;
 using Messaging.Domain.OutgoingMessages;
 using Messaging.Domain.Transactions.MoveIn;
 using Messaging.Infrastructure.Configuration.DataAccess;
